requirements.txt
asgiref==3.3.4
Django==3.2.3
Django-Bootstrap3-Validator==0.3.3
django-crispy-forms==1.11.2
flake8==3.9.2
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
